+++ winter-androidx-fragment/src/main/java/io/jentz/winter/androidx/fragment/exportAndroidTypes.kt
@@ -1,0 +1,41 @@
+package io.jentz.winter.androidx.fragment
+
+import androidx.activity.ComponentActivity
+import androidx.activity.OnBackPressedDispatcherOwner
+import androidx.fragment.app.FragmentActivity
+import androidx.fragment.app.FragmentManager
+import androidx.savedstate.SavedStateRegistryOwner
+import io.jentz.winter.Component
+
+internal fun exportAndroidTypes(
+    instance: Any,
+    enableWinterFragmentFactory: Boolean,
+    builder: Component.Builder
+) {
+    with(builder) {
+        if (instance is SavedStateRegistryOwner) {
+            constant(instance)
+            constant(instance.savedStateRegistry)
+        }
+
+        if (instance is OnBackPressedDispatcherOwner) {
+            constant(instance)
+            constant(instance.onBackPressedDispatcher)
+        }
+
+        if (instance is ComponentActivity) {
+            constant(instance)
+        }
+
+        if (instance is FragmentActivity) {
+            constant(instance)
+            constant(instance.supportFragmentManager)
+
+            if (enableWinterFragmentFactory) {
+                eagerSingleton(
+                    onPostConstruct = { instance<FragmentManager>().fragmentFactory = it }
+                ) { WinterFragmentFactory(this) }
+            }
+        }
+    }
+}