+++ CHANGELOG.md
@@ -8,6 +8,8 @@ Version 0.10.0
 * Remove fragment support from winter-androidx injection adapters
 * Add support for AndroidX view model injection to winter-androidx
 * Add winter-androidx-viewmodel-savedstate module for support for view models with saved state handle
+* Add winter-androidx-fragment module with injection adapters with Fragment support
+* Add WinterFragmentFactory (again) for constructor injection support in fragments
 * Add Component.Builder#containsKey method to check if a key was already registered
 * Remove check if target key exists from Component.Builder#alias